STYLE: More style fixes
Adds consistency in the use of vcl vs. vnl vs. std functions. Also strips
out some redundant includes in Utilities.

+++ Applications/ComputeImageQuantiles/ComputeImageQuantiles.cxx
@@ -164,7 +164,7 @@ void writeQuantilesToJSONFile( const std::vector<float>& quantiles,
     boost::property_tree::ptree quantilesJSON;
     boost::property_tree::ptree quantileValuesJSON;
 
-    for( size_t i=0; i<quantiles.size(); ++i )
+    for( unsigned int i=0; i<quantiles.size(); ++i )
       {
       boost::property_tree::ptree quantileElementJSON;
       boost::property_tree::ptree quantileValueElementJSON;
